Adding fix for broken spec
The issue was that the specs were trying to assert that the git author
was not set.  However, even though it was unset in the spec repo dir, it
*was* set globally in the `~/.gitconfig`.  This resulted in failing
specs.  The solution was to stub out the `duet.env` namespace in the
context of the specs.  It was necessary to do this with an environmental
variable instead of stubbing because the actually `git` run in the specs
(by shelling out) use the real executables and lib for this gem (in the
case of integration specs), which are not affected by the stubs.